backend/fastapi/requirements.txt
anyio==3.6.1
click==8.1.3
colorama==0.4.5
fastapi==0.85.0
greenlet==1.1.3
h11==0.13.0
httptools==0.5.0
idna==3.4
Pillow==9.2.0
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
